update complate project status
