+++ public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/js/app.js": "/js/app.js?id=10ed06894281c002d5624e50d38af6a4",
+    "/js/app.js": "/js/app.js?id=ead2419476415170e059fb29fa37fd3d",
     "/css/app.css": "/css/app.css?id=e1d1afead32875a111cbad54fb1ea29e"
 }
