--- resources/js/Pages/Package/Partials/Overview.vue
@@ -1,247 +1,0 @@
-<template>
-    <div class="row">
-        <div class="col-md-12 col-xl-8 col-12">
-            <div class="row">
-                <div class="col-12 mb-2">
-                    <div class="card">
-                        <div class="card-header">
-                            <h4 class="mb-0">Project Summary</h4>
-                        </div>
-                        <div class="card-body">
-                            <p class="newlineStringStyle">{{ info.description }}</p>
-
-                            <ul class="list-group list-group-flush">
-                                <li class="list-group-item px-0">
-                                    <div class="d-flex justify-content-between align-items-center">
-                                        <div class="d-flex align-items-center">
-                                            <vue-feather type="calendar"/>
-                                            <div class="ms-2">
-                                                <h5 class="mb-0 text-body">Start Date</h5>
-                                            </div>
-                                        </div>
-                                        <div>
-                                            <div>
-                                                <p class="text-dark mb-0 fw-semibold">{{ moment(info.start).format('D MMM  YYYY') }}</p>
-                                            </div>
-                                        </div>
-                                    </div>
-                                </li>
-                                <li class="list-group-item px-0">
-                                    <div class="d-flex justify-content-between align-items-center">
-                                        <div class="d-flex align-items-center">
-                                            <vue-feather type="calendar"/>
-                                            <div class="ms-2">
-                                                <h5 class="mb-0 text-body">End Date</h5>
-                                            </div>
-                                        </div>
-                                        <div>
-                                            <div>
-                                                <p class="text-dark mb-0 fw-semibold">{{ moment(info.end).format('D MMM  YYYY') }}</p>
-                                            </div>
-                                        </div>
-                                    </div>
-                                </li>
-                                <li class="list-group-item  px-0">
-                                    <div class="d-flex justify-content-between align-items-center">
-                                        <div class="d-flex align-items-center">
-                                            <vue-feather type="clock"/>
-                                            <div class="ms-2">
-                                                <h5 class="mb-0 text-body">Estimate
-                                                    Time</h5>
-                                            </div>
-                                        </div>
-                                        <div>
-                                            <div>
-                                                <p class="text-dark mb-0 fw-semibold">{{ estimateTimes }}Days</p>
-                                            </div>
-                                        </div>
-                                    </div>
-                                </li>
-                                <li class="list-group-item px-0 pb-0" v-if="props.info.invoice">
-                                    <div class="d-flex justify-content-between align-items-center">
-                                        <div class="d-flex align-items-center">
-                                            <vue-feather type="dollar-sign"/>
-                                            <div class="ms-2">
-                                                <h5 class="mb-0 text-body">Cost</h5>
-                                            </div>
-                                        </div>
-                                        <div>
-                                            <div>
-                                                <p class="text-dark mb-0 fw-semibold">{{ info.invoice.grand_total }} Tk</p>
-                                            </div>
-                                        </div>
-                                    </div>
-                                </li>
-
-                            </ul>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-12 mb-2">
-                    <div class="card">
-                        <div class="card-header">
-                            <h4 class="mb-0">Project Credintials</h4>
-                            <span class="cursor-pointer" @click="viewCredintials">
-                                                        <vue-feather :type="showCredintials ? 'eye' : 'eye-off'"/>
-                                                    </span>
-                        </div>
-                        <div class="card-body" v-if="showCredintials">
-                            <p class="newlineStringStyle">{{ info.credential }}</p>
-                        </div>
-                    </div>
-                </div>
-
-                <div class="col-md-12 mb-2" v-if="props.info.invoice">
-                    <!-- card -->
-                    <div class="card">
-                        <div class="card-header">
-                            <h4 class="mb-0">Budget </h4>
-                        </div>
-                        <div class="card-body">
-                            <div class="row">
-                                <!-- col -->
-                                <div class="col-lg-4 col-md-12 col-12 bg-">
-                                    <div class="d-flex align-items-center justify-content-between">
-                                        <div>
-                                            <h2 class="h1 fw-bold mb-0">{{ props.info.invoice?.grand_total ?? 0}}</h2>
-                                            <p class="mb-0">Total Budget</p>
-                                        </div>
-                                        <span style="padding: 1rem 1.1rem" class="bg-light-warning rounded-circle d-flex align-items-center justify-content-center">Tk</span>
-                                    </div>
-                                </div>
-                                <div class="col-lg-4 col-md-12 col-12 bg-">
-                                    <div class="d-flex align-items-center justify-content-between">
-                                        <div>
-                                            <h2 class="h1 fw-bold mb-0">{{ props.info.invoice?.pay ?? 0}}</h2>
-                                            <p class="mb-0">Total Pay</p>
-                                        </div>
-                                        <span style="padding: 1rem 1.1rem" class="bg-light-success fw-bold rounded-circle d-flex align-items-center justify-content-center">Tk</span>
-
-                                    </div>
-                                </div>
-                                <div class="col-lg-4 col-md-12 col-12 bg-">
-                                    <div class="d-flex align-items-center justify-content-between">
-                                        <div>
-                                            <h2 class="h1 fw-bold mb-0">{{ props.info.invoice?.due ?? 0}}</h2>
-                                            <p class="mb-0">Due Payment</p>
-                                        </div>
-                                        <span style="padding: 1rem 1.1rem" class="bg-light-danger fw-bold rounded-circle d-flex align-items-center justify-content-center">Tk</span>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                        <!-- row -->
-
-                    </div>
-                </div>
-                <div class="col-md-12 mb-2" v-if="props.info.invoice">
-                    <!-- card -->
-                    <div class="card">
-                        <!-- card body -->
-
-                        <div class="card-header">
-                            <h4 class="mb-0">Transaction History</h4>
-                        </div>
-                        <!-- table -->
-                        <div class="table-responsive overflow-y-hidden">
-                            <table class=" table table-striped table-borderless mb-0 text-nowrap table-hover table-centered">
-                                <thead>
-                                <tr>
-                                    <th class="bg-white">Payment Id</th>
-                                    <th class="bg-white">Taken By</th>
-                                    <th class="bg-white">Transaction Date</th>
-                                    <th class="bg-white">Amount</th>
-                                </tr>
-                                </thead>
-                                <tbody>
-                                <tr v-for="(item, index) in props.info.invoice?.transactions" :key="item.id">
-                                    <td>{{ 'Tran_#'+item.transaction_id }}</td>
-                                    <td>{{ item.received_by?.name }}</td>
-                                    <td>{{ moment(item.payment_date).format('ll') }}</td>
-                                    <td>{{ item.amount }}</td>
-                                </tr>
-                                </tbody>
-                            </table>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="col-md-12 col-xl-4 col-12">
-            <!-- card  -->
-            <div class="card mb-4 bg-primary border-primary">
-                <!-- card body  -->
-                <div class="card-body">
-
-                    <h4 class="card-title text-white">Launch Date </h4>
-                    <!-- dropdown  -->
-                    <div class="d-flex justify-content-between align-items-center mt-8">
-                        <div>
-                            <h1 class="display-4 text-white mb-1">{{ estimateTimes }} Days</h1>
-                            <p class="mb-0 text-white">{{ launchDate }}</p>
-                        </div>
-                        <div>
-                            <i class="fe fe-flag fs-1 text-light-primary"></i>
-                        </div>
-                    </div>
-
-                </div>
-
-            </div>
-            <!-- card  -->
-            <div class="card mb-4">
-                <!-- card body  -->
-                <div class="card-body">
-                    <h4 class=" card-title ">Overall Progress </h4>
-                    <ProgressChart :progress="20"/>
-                </div>
-            </div>
-            <div class="card">
-                <!-- Card header -->
-                <div class="card-header card-header-height d-flex justify-content-between align-items-center">
-                    <div>
-                        <h4 class="mb-0">Recent Activity
-                        </h4>
-                    </div>
-                    <div><a href="#">View All</a></div>
-                </div>
-                <!-- Card body -->
-                <div class="card-body">
-                    <!-- List group -->
-                    coming soon....
-                </div>
-            </div>
-        </div>
-    </div>
-</template>
-
-<script setup>
-
-import {computed, ref} from "vue";
-import moment from "moment";
-import ProgressChart from "../../../components/ProgressChart.vue";
-
-
-const props = defineProps({
-        info:{
-            type:Object,
-            required:true,
-        }
-    })
-
-    const estimateTimes = computed(()=>{
-        var a = moment(props.info.start);
-        var b = moment(props.info.end);
-        return b.diff(a, 'days') // 1
-    })
-
-    const launchDate  = computed(()=> moment(props.info.start).add(1,'days').format('LL , dddd'))
-
-    const showCredintials = ref(false)
-    const viewCredintials =()=> showCredintials.value = !showCredintials.value;
-
-</script>
-
-<style scoped>
-
-</style>
